Renamed files FeatureNNInteraction.h and NNInteractionReadWrite.h into more consisten FeatureContactInteractionSc.h and FeatureContactInteractionReadWrite.h . The latter will also be used in FeatureContactInteractionBcc
